Split `matmul` implementation by responsibility
