+++ contrib/standalone_buffer/Makefile
@@ -66,6 +66,9 @@ $(SETUPROOT)/BufferExpander.o: $(SETUPROOT)/solvcon/buffer/BufferExpander.cpp Ma
 $(SETUPROOT)/SimpleArray.o: $(SETUPROOT)/solvcon/buffer/SimpleArray.cpp Makefile
 	c++ $(CFLAGS) -c $< -o $@
 
+$(SETUPROOT)/matmul.o: $(SETUPROOT)/solvcon/buffer/matmul.cpp Makefile
+	c++ $(CFLAGS) -c $< -o $@
+
 $(SETUPROOT)/blas_compat.o: $(SETUPROOT)/solvcon/math/blas_compat.cpp Makefile
 	c++ $(CFLAGS) -c $< -o $@
 
@@ -81,6 +84,7 @@ $(SETUPROOT)/radix_tree.o: $(SETUPROOT)/solvcon/profiling/RadixTree.cpp Makefile
 OBJS = \
 	$(SETUPROOT)/BufferExpander.o \
 	$(SETUPROOT)/SimpleArray.o \
+	$(SETUPROOT)/matmul.o \
 	$(SETUPROOT)/blas_compat.o \
 	$(SETUPROOT)/wrap_ConcreteBuffer.o \
 	$(SETUPROOT)/wrap_SimpleArray.o \
