+++ cpp/solvcon/buffer/CMakeLists.txt
@@ -14,11 +14,14 @@ set(SOLVCON_BUFFER_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/SimpleCollector.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/loop.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/matmul.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/matmul_executor.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/matmul_plan.hpp
     CACHE FILEPATH "" FORCE)
 
 set(SOLVCON_BUFFER_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/BufferExpander.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/SimpleArray.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/matmul.cpp
     CACHE FILEPATH "" FORCE)
 
 set(SOLVCON_BUFFER_PYMODHEADERS
